requirements-production.txt
# Production Requirements - Steam Indie Analytics
# 本番環境用最小構成（開発・テスト依存関係を除外）

# Core Python Dependencies
python-dotenv>=1.0.0          # 環境変数管理
requests>=2.28.0              # HTTP リクエスト（Steam API連携）
aiohttp>=3.8.0                # 非同期HTTPリクエスト
schedule>=1.2.0               # 定期実行・バッチ処理
tenacity>=8.0.0               # リトライ機能（Steam APIレート制限対応）

# Data Processing & Database
pandas>=2.0.0                 # データ操作・変換
numpy>=1.24.0                 # 数値計算
sqlalchemy>=2.0.0            # ORM・データベース操作
psycopg2-binary>=2.9.0        # PostgreSQL接続ドライバ
pydantic>=2.0.0               # データバリデーション・型安全性

# Redis & Caching
redis==5.0.1                  # Redis クライアント
hiredis==2.2.3                # Redis 高速パーサー

# Data Analysis & Statistics
scipy==1.11.4                 # 統計処理・科学計算
scikit-learn==1.3.2           # 機械学習・前処理
statsmodels==0.14.0           # 統計モデリング

# Web Scraping & APIs
beautifulsoup4==4.12.2        # HTMLパーシング
aiofiles==23.2.1              # 非同期ファイルI/O

# Visualization & Dashboard
streamlit>=1.29.0             # ダッシュボード・Webアプリ
plotly>=5.15.0                # インタラクティブ可視化
seaborn>=0.12.0               # 統計可視化
matplotlib>=3.7.0             # 基本的なグラフ作成

# Logging & Monitoring
loguru==0.7.2                 # 高機能ログ出力
tqdm==4.66.1                  # プログレスバー
prometheus-client==0.19.0     # メトリクス収集

# Date & Time
pytz==2023.3                  # タイムゾーン処理
python-dateutil==2.8.2       # 日付解析・操作

# Utilities
click==8.1.7                  # CLI構築
rich==13.7.0                  # リッチなコンソール出力

# AI & LLM APIs
google-generativeai>=0.8.0    # Gemini API（分析コメント生成）

# Firebase/Firestore
google-cloud-firestore>=2.13.0  # Firestore NoSQLデータベース
firebase-admin>=6.4.0           # Firebase Admin SDK